smarter_school/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
psycopg2-binary>=2.9
python-dotenv>=1.0.0
pytest-django>=4.5
dj-database-url
gunicorn>=21.2.0
django-cors-headers>=4.3.1
